Remove usage of `login.dfe.async-retry` from "Update and refresh" audit function.
This is not an API endpoint of a middle-tier component and does not fit within the `login.dfe.api-client` pattern.

It seems unlikely that this call would need to be rewired to a .NET equivalent API in the future since the API is actually a LogicApp endpoint. Unless that were to change; the future .NET equivalent of the Manage component would theortically just need to make the same POST request to the same LogicApp.